Move PhysicalNode from plan folder to rel folder
