+++ core/src/main/java/org/apache/calcite/adapter/enumerable/EnumerableRel.java
@@ -18,8 +18,8 @@ package org.apache.calcite.adapter.enumerable;
 
 import org.apache.calcite.linq4j.tree.BlockStatement;
 import org.apache.calcite.plan.DeriveMode;
-import org.apache.calcite.plan.PhysicalNode;
 import org.apache.calcite.plan.RelTraitSet;
+import org.apache.calcite.rel.PhysicalNode;
 import org.apache.calcite.util.Pair;
 
 import java.util.List;
