+++ core/src/main/java/org/apache/calcite/plan/volcano/RelSet.java
@@ -17,13 +17,13 @@
 package org.apache.calcite.plan.volcano;
 
 import org.apache.calcite.plan.Convention;
-import org.apache.calcite.plan.PhysicalNode;
 import org.apache.calcite.plan.RelOptCluster;
 import org.apache.calcite.plan.RelOptListener;
 import org.apache.calcite.plan.RelOptUtil;
 import org.apache.calcite.plan.RelTrait;
 import org.apache.calcite.plan.RelTraitDef;
 import org.apache.calcite.plan.RelTraitSet;
+import org.apache.calcite.rel.PhysicalNode;
 import org.apache.calcite.rel.RelNode;
 import org.apache.calcite.rel.convert.Converter;
 import org.apache.calcite.rel.core.CorrelationId;
